Release 0.2.0 docs, migration guide, and build hygiene
Bump the version to 0.2.0 everywhere (pyproject, CMake, package, Sphinx,
Doxygen) and describe the project as science-first rather than
high-performance. Add docs/sphinx/migration.rst with every retired
spelling and its replacement, CONTRIBUTING.md with the ground rules and
development loop, and refresh getting_started, the API index, the
readme and the notes (FOOTPRINT, GAP_ANALYSIS, SOLVER_CONTRACTS) for
the one-Result / solve_until / tiered-namespace API. Sphinx now imports
the compiled extension instead of mocking it, so native signatures and
docstrings render; the few docstrings docutils misread are rewritten.

Build hygiene: a biotransport_warnings INTERFACE target carries a strict
warning set and is linked privately by the library, the extension and
the tests (BIOTRANSPORT_WERROR=ON makes them errors; the native Debug CI
job builds that way). The cheap warnings it surfaced are fixed: the
scalar solvers now name OUTWARD_FLUX explicitly in their boundary
switches, IonSpecies constructor parameters no longer shadow members,
and signed/unsigned and float/double mixes are cast explicitly. The
pytest configuration moves into pyproject.toml, the CI lint job runs
the pre-commit hooks (whose formatting fixes are applied here), and the
unreferenced dev.sh, Dockerfile, docker-compose.yml and environment.yml
are removed.

+++ cpp/src/physics/mass_transport/gray_scott.cpp
@@ -197,7 +197,7 @@ GrayScottRunResult GrayScottSolver::simulate(const std::vector<float>& u0,
             }
             last_check_u = u;
             last_check_v = v;
-            stable_this_check = stable_tol > 0.0 && max_diff < stable_tol;
+            stable_this_check = stable_tol > 0.0 && static_cast<double>(max_diff) < stable_tol;
         }
 
         if (step % steps_between_frames == 0 || step == total_steps) {
